+++ src/utils/index.js
@@ -1,0 +1,10 @@
+export function a11yProps(index) {
+  return {
+    id: `simple-tab-${index}`,
+    "aria-controls": `simple-tabpanel-${index}`,
+  };
+}
+
+export function isDark(theme) {
+  return theme === "dark";
+}
